refactor and remove unused cart models
