--- storefront/modules/cart/models/AddToCartModel.ts
@@ -1,5 +1,0 @@
-export type AddToCartModel = {
-  productId: number;
-  quantity: number;
-  parentProductId?: number | null;
-};
