--- storefront/modules/cart/models/Cart.ts
@@ -1,7 +1,0 @@
-import { CartItem } from './CartItem';
-
-export type Cart = {
-  id: number;
-  customerId: string;
-  cartDetails: CartItem[];
-};
