--- storefront/modules/cart/models/CartItem.ts
@@ -1,5 +1,0 @@
-export type CartItem = {
-  id: number;
-  productId: number;
-  quantity: number;
-};
