--- storefront/modules/cart/models/UpdateCartModel.ts
@@ -1,7 +1,0 @@
-export type UpdateCartModel = {
-  cartItemId: number;
-  userId: string;
-  productId: number;
-  quantity: number;
-  status: string;
-};
